backend/pyproject.toml
[project]
name = "vas-dj-saas-project"
version = "0.1.0"
description = ""
authors = [
    {name = "vibbsdod",email = "vibbs.dod@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
package-mode = false
dependencies = [
    "django (>=5.2.4,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "celery (>=5.5.3,<6.0.0)",
    "redis (>=6.2.0,<7.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "boto3 (>=1.39.14,<2.0.0)",
    "gunicorn (>=23.0.0,<24.0.0)",
    "drf-spectacular (>=0.27.0,<1.0.0)",
    "stripe (>=11.3.0,<12.0.0)",
    "jinja2 (>=3.1.0,<4.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=3.0.1,<4.0.0)",
    "python-json-logger (>=3.3.0,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "djangorestframework-camel-case (>=1.4.2,<2.0.0)",
    "django-cors-headers (>=4.3.0,<5.0.0)",
    "django-ratelimit (>=4.1.0,<5.0.0)",
    "django-redis (>=5.4.0,<6.0.0)",
    "drf-nested-routers (>=0.94.1,<1.0.0)",
    "sentry-sdk[django] (>=2.0.0,<3.0.0)",
    # Observability dependencies
    "opentelemetry-api (>=1.27.0,<2.0.0)",
    "opentelemetry-sdk (>=1.27.0,<2.0.0)",
    "opentelemetry-instrumentation (>=0.48b0,<1.0.0)",
    "opentelemetry-instrumentation-django (>=0.48b0,<1.0.0)",
    "opentelemetry-instrumentation-celery (>=0.48b0,<1.0.0)",
    "opentelemetry-instrumentation-psycopg2 (>=0.48b0,<1.0.0)",
    "opentelemetry-instrumentation-redis (>=0.48b0,<1.0.0)",
    "opentelemetry-instrumentation-requests (>=0.48b0,<1.0.0)",
    "opentelemetry-exporter-prometheus (>=0.48b0,<1.0.0)",
    "prometheus-client (>=0.21.0,<1.0.0)",
    "django-prometheus (>=2.3.1,<3.0.0)",
    # Testing dependencies
    "pytest (>=8.0.0,<9.0.0)",
    "pytest-django (>=4.8.0,<5.0.0)",
    "pytest-cov (>=4.0.0,<5.0.0)",
    "pytest-xdist (>=3.5.0,<4.0.0)",
    "factory-boy (>=3.3.0,<4.0.0)",
    "freezegun (>=1.4.0,<2.0.0)",
    # Code Quality & Security
    "black (>=24.0.0,<25.0.0)",
    "isort (>=5.13.0,<6.0.0)",
    "ruff (>=0.6.0,<1.0.0)",
    "mypy (>=1.11.0,<2.0.0)",
    "bandit[toml] (>=1.7.0,<2.0.0)",
    "safety (>=3.2.0,<4.0.0)",
    "pre-commit (>=3.8.0,<4.0.0)",
    # Additional type stubs
    "django-stubs[compatible-mypy] (>=5.1.0,<6.0.0)",
    "djangorestframework-stubs[compatible-mypy] (>=3.15.0,<4.0.0)",
    "types-redis (>=4.6.0,<5.0.0)",
    "types-requests (>=2.32.0,<3.0.0)",
]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
exclude = '''
/(
    \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_django = "django"
known_first_party = ["apps", "config"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "DJANGO", "FIRSTPARTY", "LOCALFOLDER"]
skip_glob = ["*/migrations/*"]

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
    "ARG001", # unused-function-args
    "SIM118", # key-in-dict
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191",  # indentation contains tabs
]
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
    "*/migrations/*",
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"test_*.py" = ["ARG001"]
"**/tests/*" = ["ARG001"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
ignore_missing_imports = true
warn_unused_ignores = true
warn_redundant_casts = true
warn_unused_configs = true
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true

[tool.django-stubs]
django_settings_module = "config.settings.development"

[tool.bandit]
exclude_dirs = ["tests", "*/migrations/*", "venv", ".venv"]
skips = ["B101", "B601"]

[tool.bandit.assert_used]
skips = ["**/test_*.py", "**/tests/*"]

[tool.coverage.run]
source = "."
omit = [
    "*/migrations/*",
    "*/venv/*",
    "*/env/*",
    "manage.py",
    "config/wsgi.py",
    "config/asgi.py",
    "*/tests/*",
    "conftest.py",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.html]
directory = "htmlcov"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
